requirements.txt
# fastapi
# uvicorn
# sentence-transformers
# scikit-learn
# pypdf
# python-multipart
# transformers
# torch


fastapi
uvicorn
sentence-transformers
scikit-learn
pypdf
python-multipart
transformers
torch
groq
python-dotenv
pytesseract
pdf2image
pillow
endee